% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gist.R
\name{create_gist}
\alias{create_gist}
\title{Create a gist}
\usage{
create_gist(
  files,
  description = NULL,
  public = TRUE,
  github_pat = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{files}{A named list where names are the desired filenames and the values are the string
content of those files.}

\item{description}{A brief, optional description of the gist (character string).}

\item{public}{A logical flag indicating whether the gist is public (`TRUE`) or secret (`FALSE`).
Default is `TRUE`.}

\item{github_pat}{Your GitHub Personal Access Token (PAT). It is recommended to set this as an
environment variable (e.g., 'GITHUB_PAT').}
}
\value{
A list containing the parsed JSON response from the GitHub API, which includes the Gist
 `id`, `url`, and `files` information.
}
\description{
Allows you to add a new gist with one or more files.
}
\keyword{internal}
