% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_gist.R
\name{delete_gist}
\alias{delete_gist}
\title{Delete a gist}
\usage{
delete_gist(gist_id, github_pat = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{gist_id}{The unique identifier of the gist (character string).}

\item{github_pat}{Your GitHub Personal Access Token (PAT). It is recommended to set this as an
environment variable (e.g., 'GITHUB_PAT').}
}
\value{
No return value. Called for its side effects.
}
\description{
Allows you to delete a gist.
}
\keyword{internal}
