% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gist.R
\name{get_gist}
\alias{get_gist}
\title{Get a gist}
\usage{
get_gist(gist_id)
}
\arguments{
\item{gist_id}{The unique identifier of the gist (character string).}
}
\value{
A list containing the parsed JSON response from the GitHub API, which includes the Gist
 `id`, `url`, and `files` information.
}
\description{
Allows you to get a gist.
}
\keyword{internal}
