% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{Join a Live Coding Session}
\usage{
join(session_id, refresh_rate = 5, stop_after = 360)
}
\arguments{
\item{session_id}{The ID of the session being broadcasted. This is provided by the broadcaster.}

\item{refresh_rate}{Time in seconds between refreshes. Default is `5`.}

\item{stop_after}{Number of refreshes after which to automatically terminate. NULL for infinite
loop. Default is `360` (30 minutes of running).}
}
\value{
No return value. Called for its side effects.
}
\description{
Connects your RStudio session to a broadcaster's session. This allows you to follow along with
code changes in real-time.
}
