% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecode.R
\name{livecode}
\alias{livecode}
\title{Live Code}
\usage{
livecode(refresh_rate = 5, stop_after = 360, clean_up = TRUE)
}
\arguments{
\item{refresh_rate}{Time in seconds between refreshes. Default is `5`.}

\item{stop_after}{Number of refreshes after which to automatically terminate. NULL for infinite
loop. Default is `360` (30 minutes of running).}

\item{clean_up}{Whether to delete the gist when the livecoding ends.}
}
\description{
Continuously monitors the active RStudio source file and pushes changes to a GitHub Gist for
live broadcasting.
}
