% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecode.R
\name{livecode_script}
\alias{livecode_script}
\title{Live Code Script}
\usage{
livecode_script(
  context_id = NULL,
  refresh_rate = 5,
  stop_after = 360,
  clean_up = TRUE
)
}
\arguments{
\item{context_id}{The ID of an rstudioapi's `document_context` to broadcast. If `NULL`, it will
load the currently open file. Default is `NULL`.}

\item{refresh_rate}{Time in seconds between refreshes. Default is `5`.}

\item{stop_after}{Number of refreshes after which to automatically terminate. NULL for infinite
loop. Default is `360` (30 minutes of running).}

\item{clean_up}{Whether to delete the gist when the livecoding ends.}
}
\value{
No return value. Called for its side effects.
}
\description{
Continuously monitors the active RStudio source file and pushes changes to a GitHub Gist for
live broadcasting.
}
