% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecoding-package.R
\docType{package}
\name{livecoding-package}
\alias{livecoding}
\alias{livecoding-package}
\title{Livecoding for R.}
\description{
Facilitates real-time live coding sessions within RStudio by broadcasting code changes to a GitHub Gist and allowing other users to join and follow along.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://myownrobs.github.io/myownrobs/Estacion-R/estacion-r-streaming/livecoding/}
  \item Report bugs at \url{https://github.com/Estacion-R/estacion-r-streaming/issues}
}

}
\author{
\strong{Maintainer}: Juan Cruz Rodriguez \email{jcrodriguez@unc.edu.ar}

}
